Fixed model

+++ src/models/pet.cr
@@ -2,7 +2,7 @@ class Pet < Granite::Base
   connection mysql
   table pets
 
-  column id : Int64, primary: true
+  column id : Int32, primary: true
   column name : String?
   column breed : String?
   column age : Int32?
